refactor(entities): expose product selectors and use slice entrypoints
