+++ src/entities/product/index.js
@@ -1,4 +1,5 @@
 export * from './api/client';
+export * from './model/selectors';
 export * as productSelectors from './model/selectors';
 export * from './model/useProduct';
 export { default as ProductCard } from './ui/ProductCard/ProductCard';
