+++ src/widgets/products-catalog/model/useProductsCatalog.js
@@ -4,7 +4,7 @@ import { useRouter, useSearchParams } from 'next/navigation';
 import { useEffect, useMemo, useRef, useState, useTransition } from 'react';
 
 import { getBrands } from '@/entities/brand';
-import { getProducts } from '@/entities/product/api/client';
+import { getProducts } from '@/entities/product';
 
 export function useProductsCatalog() {
   const router = useRouter();
